requirements.txt
a2a-sdk
fastapi
uvicorn
openai
sse-starlette
httpx
mcp
flask